+++ src/routes/Movies.tsx
@@ -1,32 +1,74 @@
-import axios from "axios";
-import { Link, useLoaderData } from "react-router-dom";
-
-export type MoviesType = {
-  id: number;
-  title: string;
-  content: string;
-};
+import { useState } from "react";
+import { API } from "../components/API";
+import { Transition } from "@headlessui/react";
+import clsx from "clsx";
+import Moviespagebutton from "../components/MoviesPageButton";
+import { Movie } from "../types/api";
+import { useLoaderData } from "react-router-dom";
 
 export async function MoviesLoader() {
-  const result = await axios.get<MoviesType[]>("http://localhost:3000/movies");
-  return result.data;
+  return API.fourByFour();
 }
-
 export function Movies() {
-  const movies = useLoaderData() as MoviesType[];
+  const API = useLoaderData() as Movie[][];
+  const [number, setNumber] = useState(0);
+  const [previousvalue, setpreviousvalue] = useState(0);
   return (
-    <div>
-      <div className="flex flex-col gap-3">
-        {movies.map((movie) => (
-          <Link
-            className="underline text-indigo-600"
-            key={movie.id}
-            to={`/movie/${movie.id}`}
-          >
-            {movie.title}
-          </Link>
-        ))}
-      </div>
+    <div className="relative bg-dark">
+      {API.map((card, index) => {
+        return (
+          <div className="flex flex-col">
+            <div className=" flex flex-col items-center">
+              <div className="h-28"></div>
+              <div className="absolute top-4">
+                <Transition
+                  show={number == index}
+                  enter="transition transform translate-x duration-[750ms]"
+                  enterFrom={clsx(
+                    "opacity-5 scale-50",
+                    index < previousvalue ? "-translate-x-60" : "translate-x-60"
+                  )}
+                  enterTo="opacity-100 scale-100 translate-x-0"
+                  leave="transition transform duration-[750ms]"
+                  leaveFrom=""
+                  leaveTo={clsx(
+                    "opacity-5 scale-50",
+                    number < previousvalue
+                      ? "translate-x-60"
+                      : "-translate-x-60"
+                  )}
+                  beforeEnter={() => {
+                    console.log(
+                      "Enter",
+                      index < previousvalue ? "Decreasing" : "Increasing"
+                    );
+                  }}
+                  afterEnter={() => {
+                    setpreviousvalue(index);
+                  }}
+                  afterLeave={() => {
+                    setpreviousvalue(index);
+                  }}
+                >
+                  <div className="grid grid-rows-2 grid-cols-2 gap-x-5 gap-y-5">
+                    {card.map((card) => {
+                      return (
+                        <img
+                          src={`https://image.tmdb.org/t/p/w500${card?.poster_path}`}
+                          alt=""
+                          className="w-40 h-60 rounded-lg flex-grow-0"
+                          key={card?.id}
+                        />
+                      );
+                    })}
+                  </div>
+                </Transition>
+              </div>
+            </div>
+          </div>
+        );
+      })}
+      <Moviespagebutton number={number} setNumber={setNumber} />
     </div>
   );
 }
